feat: hide selector if there are no mutations
